backend/chat/serializers.py
from rest_framework import serializers
from .models import Room, Message

class RoomSerializer(serializers.ModelSerializer):
    created_by_username = serializers.CharField(source='created_by.username', read_only=True)

    class Meta:
        model = Room
        fields = ['id', 'name', 'created_by', 'created_by_username', 'created_at']
        read_only_fields = ['created_by', 'created_at']


class MessageSerializer(serializers.ModelSerializer):
    user_username = serializers.CharField(source='user.username', read_only=True)

    class Meta:
        model = Message
        fields = ['id', 'room', 'user', 'user_username', 'text', 'created_at']
        read_only_fields = ['user', 'created_at']
