backend/requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.6.0
drf-nested-routers==0.94.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3

# 👇 добавляем для realtime
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
